# docs/swagger.yaml
basePath: /
definitions:
...
  request.UserUpdateRequest:
    properties:
      age:
        description: |-
          User's age (required, must be between 1 and 140).
          @json
          @jsonTag age
          @jsonExample 30
          @binding required,min=1,max=140
        example: 30
        maximum: 140
        minimum: 1
        type: integer
      name:
        description: |-
          User's name (required, minimum of 4 characters, maximum of 100 characters).
          Example: John Doe
          @json
          @jsonTag name
          @jsonExample John Doe
          @binding required,min=4,max=100
        example: John Doe
        maxLength: 100
        minLength: 4
        type: string
    required:
    - age
    - name
    type: object
...
  rest_err.Causes:
    description: Structure representing the causes of an error.
    properties:
      field:
        description: |-
          Field associated with the error cause.
          @json
          @jsonTag field
        example: name
        type: string
      message:
        description: |-
          Error message describing the cause.
          @json
          @jsonTag message
        example: name is required
        type: string
    type: object
  rest_err.RestErr:
    description: Structure for describing why the error occurred
    properties:
      causes:
        description: Error causes.
        items:
          $ref: '#/definitions/rest_err.Causes'
        type: array
      code:
        description: Error code.
        example: 500
        type: integer
      error:
        description: Error description.
        example: internal_server_error
        type: string
      message:
        description: Error message.
        example: error trying to process request
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for crud operations on users
  title: CRUD-GO | Rafael Freire
  version: "1.0"
paths:
...
  /updateUser/{userId}:
    put:
      consumes:
      - application/json
      description: Updates user details based on the ID provided as a parameter.
      parameters:
      - description: ID of the user to be updated
        in: path
        name: userId
        required: true
        type: string
      - description: User information for update
        in: body
        name: userRequest
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdateRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Update User
      tags:
      - Users
schemes:
- http
swagger: "2.0"